feat: add shared HubSpot client
