+++ src/server/search_contacts.ts
@@ -1,12 +1,10 @@
 import { createClient, type SupabaseClient } from '@supabase/supabase-js'
 import type { Database } from '../integrations/supabase/types'
-import rateLimiter from './rate_limiter_memory'
 import {
   SUPABASE_URL,
   SUPABASE_SERVICE_ROLE_KEY,
-  HUBSPOT_CLIENT_ID,
-  HUBSPOT_CLIENT_SECRET,
 } from './config'
+import { searchContacts as hubspotSearch } from '../integrations/hubspot/client'
 
 export interface ContactRecord {
   id: string
@@ -15,46 +13,6 @@ export interface ContactRecord {
 
 const supabase = createClient<Database>(SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY)
 
-async function ensureAccessToken(portal_id: string, sb: SupabaseClient<Database> = supabase): Promise<string> {
-  const { data, error } = await sb
-    .from('hubspot_tokens')
-    .select('access_token, refresh_token, expires_at')
-    .eq('portal_id', portal_id)
-    .maybeSingle()
-
-  if (error || !data) {
-    throw new Error('Token fetch failed')
-  }
-
-  if (data.expires_at && new Date(data.expires_at).getTime() > Date.now() + 60_000) {
-    return data.access_token
-  }
-
-  const resp = await fetch('https://api.hubapi.com/oauth/v1/token', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-    body: new URLSearchParams({
-      grant_type: 'refresh_token',
-      refresh_token: data.refresh_token,
-      client_id: HUBSPOT_CLIENT_ID,
-      client_secret: HUBSPOT_CLIENT_SECRET,
-    }).toString(),
-  })
-
-  if (!resp.ok) throw new Error('Refresh failed')
-  const json: any = await resp.json()
-
-  await sb
-    .from('hubspot_tokens')
-    .update({
-      access_token: json.access_token,
-      refresh_token: json.refresh_token ?? data.refresh_token,
-      expires_at: new Date(Date.now() + json.expires_in * 1000).toISOString(),
-    })
-    .eq('portal_id', portal_id)
-
-  return json.access_token
-}
 
 export async function searchLocal(
   portal_id: string,
@@ -78,31 +36,7 @@ async function searchRemote(
   sb: SupabaseClient<Database> = supabase,
   fetchFn: typeof fetch = fetch
 ): Promise<ContactRecord[]> {
-  const accessToken = await ensureAccessToken(portal_id, sb)
-  await rateLimiter.take(portal_id)
-  const response = await fetchFn('https://api.hubapi.com/crm/v3/objects/contacts/search', {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      filterGroups: [
-        {
-          filters: [
-            { propertyName: 'firstname', operator: 'CONTAINS_TOKEN', value: q },
-            { propertyName: 'lastname', operator: 'CONTAINS_TOKEN', value: q },
-            { propertyName: 'email', operator: 'CONTAINS_TOKEN', value: q },
-          ],
-        },
-      ],
-      limit: 20,
-    }),
-  })
-
-  if (!response.ok) throw new Error('HubSpot search failed')
-  const json: any = await response.json()
-  const rows: ContactRecord[] = json.results || []
+  const rows = await hubspotSearch(portal_id, q, limit, sb, fetchFn)
   const now = new Date().toISOString()
   if (rows.length) {
     await sb.from('hubspot_contacts_cache').upsert(
